+++ src/main/java/org/sourcelab/kafka/connect/apiclient/request/put/PutConnectorTopicsReset.java
@@ -1,0 +1,55 @@
+/**
+ * Copyright 2018, 2019 SourceLab.org https://github.com/SourceLabOrg/kafka-connect-client
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
+ * documentation files (the "Software"), to deal in the Software without restriction, including without limitation the
+ * rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit
+ * persons to whom the Software is furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all copies or substantial portions of the
+ * Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE
+ * WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
+ * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+package org.sourcelab.kafka.connect.apiclient.request.put;
+
+import java.io.IOException;
+import java.util.Objects;
+
+import static com.google.common.net.UrlEscapers.urlPathSegmentEscaper;
+
+/**
+ * Send a request to empty the set of active topics of a connector.
+ * https://docs.confluent.io/current/connect/references/restapi.html#put--connectors-(string-name)-topics-reset
+ */
+public class PutConnectorTopicsReset implements PutRequest<Boolean> {
+    private final String connectorName;
+
+    /**
+     * Constructor.
+     * @param connectorName Name of connector
+     */
+    public PutConnectorTopicsReset(final String connectorName) {
+        Objects.requireNonNull(connectorName);
+        this.connectorName = connectorName;
+    }
+
+    @Override
+    public String getApiEndpoint() {
+        return "/connectors/" + urlPathSegmentEscaper().escape(connectorName) + "/topics/reset";
+    }
+
+    @Override
+    public Object getRequestBody() {
+        return null;
+    }
+
+    @Override
+    public Boolean parseResponse(final String responseStr) throws IOException {
+        return true;
+    }
+}